requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.20
langchain-google-genai==1.0.3
google-cloud-monitoring==2.21.0
google-cloud-logging==3.10.0
google-cloud-pubsub==2.21.0
google-auth==2.29.0
requests==2.31.0
python-dotenv==1.0.1
PyGithub==2.3.0
